State the syslog facility keyword explicitly.
To correctly simulate the syslog format expected by pgsi, the syslog
facility keyword ('postgres') should be stated explicitly. This test case
used the name of the database ('%d'), which only worked by coincidence since
the name happened to be the same as the syslog facility keyword. This change
allows the test to work even if it is modified in the future to create and
use a different database. It also can be copied by users who wish to test
pgsi on their own database without using syslog.